fix: render a single page-level donate drawer shared by all triggers
StoryPage mounts PageSummary twice (mobile summary + desktop sticky
card), and DonateButton bundled its own Drawer + Stripe Elements tree,
so visiting a story page with ?open=donate opened two stacked drawers.

- Add donateDrawerStore, a module-level useSyncExternalStore store with
  openDonateDrawer/closeDonateDrawer, so any trigger can open the one
  drawer without threading context through server components
- Move the drawer/checkout UI into DonateDrawer, mounted once per page
  (page.tsx and embed/page.tsx); it derives projectName/pageID/presets
  from the page record, removing the duplicated fallback chain in
  PageSummary and SimplePage
- Slim DonateButton down to a trigger button for the shared drawer
- Add story-page e2e regression tests asserting ?open=donate renders
  exactly one dialog (desktop + mobile), plus an embed drawer test
- Make e2e base URL and MailPit URL overridable via E2E_BASE_URL and
  MAILPIT_URL so tests can run alongside other local apps; defaults and
  CI behavior are unchanged
