+++ playwright.config.ts
@@ -1,5 +1,10 @@
 import { defineConfig, devices } from '@playwright/test';
 
+// Port/base URL are overridable so tests can run alongside other local
+// apps occupying 3000 (e.g. E2E_BASE_URL=http://localhost:3010).
+const baseURL = process.env.E2E_BASE_URL || 'http://localhost:3000';
+const port = Number(new URL(baseURL).port) || 3000;
+
 export default defineConfig({
     testDir: './e2e',
     fullyParallel: false,
@@ -12,7 +17,7 @@ export default defineConfig({
         timeout: 10_000,
     },
     use: {
-        baseURL: 'http://localhost:3000',
+        baseURL,
         trace: 'on-first-retry',
         screenshot: 'only-on-failure',
     },
@@ -26,7 +31,8 @@ export default defineConfig({
     globalTeardown: './e2e/global-teardown.ts',
     webServer: {
         command: 'bun run build && bun run start',
-        port: 3000,
+        port,
+        env: { ...(process.env as Record<string, string>), PORT: String(port) },
         timeout: 300_000,
         reuseExistingServer: !process.env.CI,
     },
