+++ src/app/(frontend)/[pageID]/DonateDrawer.tsx
@@ -1,0 +1,682 @@
+'use client';
+
+import type { Page } from '@db/schema';
+import { presetsSchema } from '@db/schema';
+import { CheckoutForm } from '@frontend/[pageID]/CheckoutForm';
+import {
+    closeDonateDrawer,
+    openDonateDrawer,
+    useDonateDrawerOpen,
+} from '@frontend/[pageID]/donateDrawerStore';
+import { stripePromise } from '@lib/stripe/client';
+import { Elements, useElements } from '@stripe/react-stripe-js';
+import type { DonationConfig } from '@v1/donations/checkout/DonationConfig';
+import {
+    computeTipAmount,
+    estimateFee,
+    FrequencyOptions,
+    TIP_FIXED_PRESETS,
+    TIP_FIXED_THRESHOLD,
+    TIP_PERCENT_PRESETS,
+} from '@v1/donations/checkout/DonationConfig';
+import clsx from 'clsx';
+import { HappinessConfig } from 'happiness.config';
+import { useRouter, useSearchParams } from 'next/navigation';
+import { Button } from 'paris/button';
+import { Checkbox } from 'paris/checkbox';
+import { Drawer } from 'paris/drawer';
+import { Input } from 'paris/input';
+import { usePagination } from 'paris/pagination';
+import { Select } from 'paris/select';
+import { Text } from 'paris/text';
+import { TextArea } from 'paris/textarea';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import styles from 'src/app/(frontend)/[pageID]/DonateDrawer.module.scss';
+import { ButtonGroup } from 'src/components/ButtonGroup';
+import { DonationAmountSelector } from 'src/components/DonationAmountSelector';
+import { formatCurrency, formatCurrencyCompact } from 'src/util/formatCurrency';
+
+export const AmountPresets = [1000, 2500, 5000, 10000, 25000] as number[];
+
+const DrawerPages = ['details', 'payment'] as const;
+
+const STRIPE_APPEARANCE = {
+    theme: 'flat' as const,
+    labels: 'above' as const,
+    variables: {
+        borderRadius: '0px',
+        fontFamily:
+            '\'Graphik Web\', -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif',
+        colorPrimary: '#131313',
+        fontSizeBase: '14px',
+        spacingUnit: '3px',
+    },
+};
+
+const initialDonation: DonationConfig = {
+    amount: 1000,
+    frequency: 'One-time',
+    message: '',
+    coverFees: true,
+    anonymous: false,
+    projectName: HappinessConfig.name,
+    pageID: '',
+    tipPercent: 0,
+    tipFixed: 300,
+    email: '',
+    donorName: '',
+};
+
+/** Compute the total amount in cents for a given donation config */
+const computeTotal = (donation: DonationConfig) => {
+    const fee = donation.coverFees ? estimateFee(donation.amount) : 0;
+    const tip = computeTipAmount(
+        donation.amount + fee,
+        donation.tipPercent,
+        donation.tipFixed,
+    );
+    return donation.amount + fee + tip;
+};
+
+/**
+ * The single donate drawer for a page. Mounted once at the page level;
+ * opened by any DonateButton trigger (or ?open=donate) via donateDrawerStore.
+ */
+export const DonateDrawer = ({ page }: { page: Page }) => {
+    const params = useSearchParams();
+    const router = useRouter();
+    const drawerOpen = useDonateDrawerOpen();
+
+    const projectName =
+        page.fsProject ||
+        page.organizer ||
+        HappinessConfig.fiscalSponsorName ||
+        HappinessConfig.name;
+    const pageID = page.id;
+    const { presets } = page;
+
+    const firstPresetAmount = useMemo(() => {
+        if (
+            Array.isArray(presets) &&
+            presets.length > 0 &&
+            typeof presets[0]?.amount === 'number'
+        ) {
+            return presets[0].amount;
+        }
+        return 1000;
+    }, [presets]);
+
+    const [donation, setDonation] = useState<DonationConfig>({
+        ...initialDonation,
+        amount: firstPresetAmount,
+        projectName: projectName || HappinessConfig.name,
+        pageID: pageID || '',
+    });
+
+    const [showOtherAmount, setShowOtherAmount] = useState(false);
+    const [otherAmountInput, setOtherAmountInput] = useState('');
+    const [paymentLoading, setPaymentLoading] = useState(false);
+
+    const pagination = usePagination<typeof DrawerPages>('details');
+    const checkoutFormRef = useRef<HTMLFormElement>(null);
+
+    const total = useMemo(() => computeTotal(donation), [donation]);
+    const elementsMode =
+        donation.frequency === 'Monthly' ? 'subscription' : 'payment';
+
+    useEffect(() => {
+        if (params.get('open') === 'donate') {
+            openDonateDrawer();
+        }
+    }, [params]);
+
+    // Reset the shared open state when leaving the page
+    useEffect(() => closeDonateDrawer, []);
+
+    const handleClose = useCallback(() => {
+        closeDonateDrawer();
+        setDonation({
+            ...initialDonation,
+            projectName: projectName || HappinessConfig.name,
+            pageID: pageID || '',
+        });
+        setShowOtherAmount(false);
+        setOtherAmountInput('');
+        pagination.reset();
+    }, [projectName, pageID, pagination]);
+
+    const handleSuccess = useCallback(() => {
+        const name = donation.anonymous ? 'Anonymous' : 'Donor';
+        handleClose();
+        router.replace(`/${pageID}?thanks=${encodeURIComponent(name)}`);
+    }, [handleClose, donation.anonymous, pageID, router]);
+
+    const estFee = useMemo(
+        () => estimateFee(donation.amount),
+        [donation.amount],
+    );
+
+    const bottomPanel = useMemo(() => {
+        if (pagination.currentPage === 'details') {
+            return (
+                <Button
+                    onClick={() => pagination.open('payment')}
+                    disabled={donation.amount === 0}
+                    style={{ width: '100%' }}
+                >
+                    {`Continue to Payment — ${formatCurrency(total, 2)}${donation.frequency === 'One-time' ? '' : ' / month'}`}
+                </Button>
+            );
+        }
+        return (
+            <Button
+                onClick={() => {
+                    checkoutFormRef.current?.requestSubmit();
+                }}
+                disabled={paymentLoading || !donation.email}
+                style={{ width: '100%' }}
+            >
+                {paymentLoading
+                    ? 'Processing...'
+                    : `Donate ${formatCurrency(total, 2)}${donation.frequency === 'One-time' ? '' : ' / month'}`}
+            </Button>
+        );
+    }, [
+        pagination,
+        donation.amount,
+        donation.frequency,
+        donation.email,
+        total,
+        paymentLoading,
+    ]);
+
+    return (
+        <Drawer
+            title="Donate"
+            isOpen={drawerOpen}
+            onClose={() => handleClose()}
+            pagination={pagination}
+            bottomPanel={bottomPanel}
+        >
+            <Elements
+                key={`${elementsMode}-${donation.frequency}`}
+                stripe={stripePromise}
+                options={{
+                    mode: elementsMode,
+                    amount: Math.max(total, 50),
+                    currency: 'usd',
+                    appearance: STRIPE_APPEARANCE,
+                }}
+            >
+                <DonateDrawerInner
+                    donation={donation}
+                    setDonation={setDonation}
+                    showOtherAmount={showOtherAmount}
+                    setShowOtherAmount={setShowOtherAmount}
+                    otherAmountInput={otherAmountInput}
+                    setOtherAmountInput={setOtherAmountInput}
+                    total={total}
+                    estFee={estFee}
+                    onSuccess={handleSuccess}
+                    checkoutFormRef={checkoutFormRef}
+                    setPaymentLoading={setPaymentLoading}
+                    currentPage={pagination.currentPage}
+                    presets={presets}
+                />
+            </Elements>
+        </Drawer>
+    );
+};
+
+/**
+ * Inner component rendered inside <Elements>.
+ * Since the Drawer with pagination expects keyed children, but we need
+ * useStripe/useElements hooks, we render inside Elements and output
+ * the keyed page divs directly for the Drawer's pagination to pick up.
+ *
+ * NOTE: The Drawer matches children by their React `key` prop against
+ * pagination.currentPage. We return a flat array of keyed divs.
+ */
+const DonateDrawerInner = ({
+    donation,
+    setDonation,
+    showOtherAmount,
+    setShowOtherAmount,
+    otherAmountInput,
+    setOtherAmountInput,
+    total,
+    estFee,
+    onSuccess,
+    checkoutFormRef,
+    setPaymentLoading,
+    currentPage,
+    presets,
+}: {
+    donation: DonationConfig;
+    setDonation: React.Dispatch<React.SetStateAction<DonationConfig>>;
+    showOtherAmount: boolean;
+    setShowOtherAmount: (v: boolean) => void;
+    otherAmountInput: string;
+    setOtherAmountInput: (v: string) => void;
+    total: number;
+    estFee: number;
+    onSuccess: () => void;
+    checkoutFormRef: React.RefObject<HTMLFormElement | null>;
+    setPaymentLoading: (v: boolean) => void;
+    currentPage: string;
+    presets?: unknown;
+}) => {
+    const elements = useElements();
+
+    const parsedPresets = useMemo(() => {
+        const result = presetsSchema.safeParse(presets);
+        return result.success ? result.data : null;
+    }, [presets]);
+
+    const amounts = useMemo(
+        () =>
+            parsedPresets ? parsedPresets.map((p) => p.amount) : AmountPresets,
+        [parsedPresets],
+    );
+
+    const isTierView = parsedPresets?.every((p) => p.name);
+
+    const hasTip = donation.tipPercent !== 0 || donation.tipFixed !== 0;
+    const useFixedTip = donation.amount < TIP_FIXED_THRESHOLD;
+
+    // Switch tip mode when donation amount crosses the $20 threshold
+    useEffect(() => {
+        if (!hasTip) return;
+        if (useFixedTip && donation.tipFixed === 0) {
+            // Crossed below threshold — switch to fixed
+            setDonation((d) => ({
+                ...d,
+                tipFixed: TIP_FIXED_PRESETS[1],
+                tipPercent: 0,
+            }));
+        } else if (!useFixedTip && donation.tipPercent === 0) {
+            // Crossed above threshold — switch to percent
+            setDonation((d) => ({
+                ...d,
+                tipPercent: TIP_PERCENT_PRESETS[1],
+                tipFixed: 0,
+            }));
+        }
+    }, [
+        useFixedTip,
+        hasTip,
+        donation.tipFixed,
+        donation.tipPercent,
+        setDonation,
+    ]);
+
+    // Keep Elements amount in sync with donation total
+    useEffect(() => {
+        if (elements) {
+            elements.update({
+                amount: Math.max(total, 50),
+            });
+        }
+    }, [elements, total]);
+
+    const returnUrl = useMemo(
+        () =>
+            typeof window !== 'undefined'
+                ? `${window.location.origin}/v1/donations/checkout/success`
+                : '',
+        [],
+    );
+
+    return (
+        <>
+            {currentPage === 'details' && (
+                <div className="w-full flex flex-col gap-6">
+                    <ButtonGroup
+                        options={FrequencyOptions.map((f) => ({
+                            id: f,
+                            name: f,
+                        }))}
+                        selected={donation.frequency}
+                        onChange={({ id }) => {
+                            setDonation((d) => ({
+                                ...d,
+                                frequency: id as DonationConfig['frequency'],
+                            }));
+                        }}
+                    />
+                    <div className="flex flex-col gap-2">
+                        {isTierView && parsedPresets ? (
+                            <Select
+                                kind="card"
+                                label="Donation amount"
+                                hideLabel
+                                value={String(donation.amount)}
+                                onChange={(id) => {
+                                    if (id === 'other') {
+                                        setDonation((d) => ({
+                                            ...d,
+                                            amount: 0,
+                                        }));
+                                        setOtherAmountInput('');
+                                        setShowOtherAmount(true);
+                                    } else if (id) {
+                                        setDonation((d) => ({
+                                            ...d,
+                                            amount: Number(id),
+                                        }));
+                                        setShowOtherAmount(false);
+                                    }
+                                }}
+                                options={[
+                                    ...parsedPresets.map((preset) => ({
+                                        id: String(preset.amount),
+                                        node: (
+                                            <div
+                                                className="w-full flex justify-between items-center text-left"
+                                                style={{
+                                                    padding: '8px 12px',
+                                                }}
+                                            >
+                                                <div className="flex flex-col gap-0.5">
+                                                    <Text
+                                                        kind="paragraphSmall"
+                                                        weight="semibold"
+                                                    >
+                                                        {preset.name}
+                                                    </Text>
+                                                    {preset.description && (
+                                                        <Text
+                                                            kind="paragraphXSmall"
+                                                            style={{
+                                                                opacity: 0.7,
+                                                            }}
+                                                        >
+                                                            {preset.description}
+                                                        </Text>
+                                                    )}
+                                                </div>
+                                                <Text
+                                                    kind="paragraphSmall"
+                                                    weight="medium"
+                                                >
+                                                    {formatCurrencyCompact(
+                                                        preset.amount,
+                                                    )}
+                                                </Text>
+                                            </div>
+                                        ),
+                                    })),
+                                    {
+                                        id: 'other',
+                                        node: (
+                                            <div
+                                                className="w-full flex flex-col gap-0.5 text-left"
+                                                style={{
+                                                    padding: '8px 12px',
+                                                }}
+                                            >
+                                                <Text
+                                                    kind="paragraphSmall"
+                                                    weight="semibold"
+                                                >
+                                                    Other
+                                                </Text>
+                                                <Text
+                                                    kind="paragraphXSmall"
+                                                    style={{
+                                                        opacity: 0.7,
+                                                    }}
+                                                >
+                                                    Choose your own amount.
+                                                </Text>
+                                            </div>
+                                        ),
+                                    },
+                                ]}
+                            />
+                        ) : (
+                            <div className="grid grid-cols-3 gap-2">
+                                {amounts.map((amount) => (
+                                    <DonationAmountSelector
+                                        selected={donation.amount === amount}
+                                        key={amount}
+                                        onClick={() => {
+                                            setDonation((d) => ({
+                                                ...d,
+                                                amount,
+                                            }));
+                                            setShowOtherAmount(false);
+                                        }}
+                                    >
+                                        <Text
+                                            kind="paragraphSmall"
+                                            style={{ fontWeight: 500 }}
+                                        >
+                                            {formatCurrencyCompact(amount)}
+                                        </Text>
+                                    </DonationAmountSelector>
+                                ))}
+                                <DonationAmountSelector
+                                    selected={
+                                        !amounts.includes(donation.amount)
+                                    }
+                                    key="otherAmount"
+                                    onClick={() => {
+                                        setDonation((d) => ({
+                                            ...d,
+                                            amount: 0,
+                                        }));
+                                        setOtherAmountInput('');
+                                        setShowOtherAmount(true);
+                                    }}
+                                >
+                                    <Text
+                                        kind="paragraphSmall"
+                                        style={{ fontWeight: 500 }}
+                                    >
+                                        Other
+                                    </Text>
+                                </DonationAmountSelector>
+                            </div>
+                        )}
+                        <div
+                            className="overflow-clip"
+                            style={{
+                                transition: 'all 0.2s ease-in-out',
+                                maxHeight: showOtherAmount ? '40px' : 0,
+                                marginBottom: showOtherAmount ? '0px' : '-8px',
+                            }}
+                        >
+                            <Input
+                                label="Other amount"
+                                hideLabel
+                                placeholder="Enter amount"
+                                type="text"
+                                value={otherAmountInput}
+                                startEnhancer="$"
+                                onChange={(e) => {
+                                    const { value } = e.target;
+                                    if (
+                                        /^[0-9]+(\.[0-9]{0,2}|)$/.test(value) ||
+                                        value === ''
+                                    ) {
+                                        setOtherAmountInput(value);
+                                        setDonation((d) => ({
+                                            ...d,
+                                            amount: Number(value) * 100,
+                                        }));
+                                    }
+                                }}
+                            />
+                        </div>
+                    </div>
+                    <TextArea
+                        label={
+                            <div>
+                                <Text
+                                    kind="paragraphSmall"
+                                    style={{ fontWeight: 500 }}
+                                >
+                                    Message
+                                </Text>{' '}
+                                <Text kind="paragraphSmall">(optional)</Text>
+                            </div>
+                        }
+                        placeholder="Keep up the good work!"
+                        value={donation.message}
+                        style={{ minHeight: '80px' }}
+                        onChange={(e) => {
+                            const { value } = e.target;
+                            setDonation((d) => ({
+                                ...d,
+                                message: value,
+                            }));
+                        }}
+                    />
+                    <div className={clsx(styles.TempCheckboxSpacing)}>
+                        <Checkbox
+                            checked={donation.coverFees}
+                            onChange={(e) => {
+                                setDonation((d) => ({
+                                    ...d,
+                                    coverFees: Boolean(e),
+                                }));
+                            }}
+                        >
+                            <Text kind="paragraphXSmall">
+                                Add <strong>{formatCurrency(estFee, 2)}</strong>{' '}
+                                to cover processing fees
+                            </Text>
+                        </Checkbox>
+                    </div>
+                    <div className={clsx(styles.TempCheckboxSpacing)}>
+                        <Checkbox
+                            checked={donation.anonymous}
+                            onChange={(e) => {
+                                setDonation((d) => ({
+                                    ...d,
+                                    anonymous: Boolean(e),
+                                }));
+                            }}
+                        >
+                            <Text kind="paragraphXSmall">
+                                Make my donation anonymous
+                            </Text>
+                        </Checkbox>
+                    </div>
+                    <div className={clsx(styles.TempCheckboxSpacing)}>
+                        <Checkbox
+                            checked={
+                                donation.tipPercent !== 0 ||
+                                donation.tipFixed !== 0
+                            }
+                            onChange={(e) => {
+                                if (e) {
+                                    const useFixed =
+                                        donation.amount < TIP_FIXED_THRESHOLD;
+                                    setDonation((d) => ({
+                                        ...d,
+                                        tipPercent: useFixed
+                                            ? 0
+                                            : TIP_PERCENT_PRESETS[1],
+                                        tipFixed: useFixed
+                                            ? TIP_FIXED_PRESETS[1]
+                                            : 0,
+                                    }));
+                                } else {
+                                    setDonation((d) => ({
+                                        ...d,
+                                        tipPercent: 0,
+                                        tipFixed: 0,
+                                    }));
+                                }
+                            }}
+                        >
+                            <Text kind="paragraphXSmall">
+                                Add a tip for {HappinessConfig.name}
+                            </Text>
+                        </Checkbox>
+                        {donation.tipPercent !== 0 ||
+                        donation.tipFixed !== 0 ? (
+                            <div className="flex flex-col mt-4 gap-4">
+                                <Text kind="paragraphXSmall">
+                                    {HappinessConfig.tipDescription}
+                                </Text>
+                                <div className="grid grid-cols-3 gap-2">
+                                    {donation.amount < TIP_FIXED_THRESHOLD
+                                        ? TIP_FIXED_PRESETS.map((cents) => (
+                                              <DonationAmountSelector
+                                                  key={`tip-fixed-${cents}`}
+                                                  size="small"
+                                                  selected={
+                                                      donation.tipFixed ===
+                                                      cents
+                                                  }
+                                                  onClick={() => {
+                                                      setDonation((d) => ({
+                                                          ...d,
+                                                          tipFixed: cents,
+                                                          tipPercent: 0,
+                                                      }));
+                                                  }}
+                                              >
+                                                  <Text
+                                                      kind="paragraphXSmall"
+                                                      style={{
+                                                          fontWeight: 500,
+                                                      }}
+                                                  >
+                                                      {formatCurrency(cents, 0)}
+                                                  </Text>
+                                              </DonationAmountSelector>
+                                          ))
+                                        : TIP_PERCENT_PRESETS.map((pct) => (
+                                              <DonationAmountSelector
+                                                  key={`tip-pct-${pct}`}
+                                                  size="small"
+                                                  selected={
+                                                      donation.tipPercent ===
+                                                      pct
+                                                  }
+                                                  onClick={() => {
+                                                      setDonation((d) => ({
+                                                          ...d,
+                                                          tipPercent: pct,
+                                                          tipFixed: 0,
+                                                      }));
+                                                  }}
+                                              >
+                                                  <Text
+                                                      kind="paragraphXSmall"
+                                                      style={{
+                                                          fontWeight: 500,
+                                                      }}
+                                                  >
+                                                      {pct * 100}%
+                                                  </Text>
+                                              </DonationAmountSelector>
+                                          ))}
+                                </div>
+                            </div>
+                        ) : null}
+                    </div>
+                </div>
+            )}
+
+            {currentPage === 'payment' && (
+                <div className="w-full">
+                    <CheckoutForm
+                        ref={checkoutFormRef as React.Ref<HTMLFormElement>}
+                        donation={donation}
+                        onSuccess={onSuccess}
+                        returnUrl={returnUrl}
+                        onLoadingChange={setPaymentLoading}
+                        onBillingChange={(fields) =>
+                            setDonation((d) => ({ ...d, ...fields }))
+                        }
+                    />
+                </div>
+            )}
+        </>
+    );
+};
